% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binomial.p}
\alias{binomial.p}
\title{get the p-value from a binomial test where you have counts observed for some ids
and want to know if they are overrepresented among a set of "success" ids vs a background of all ids.
Example: For a set of genes of interest (GOI) each observed a number of times,
if we sample randomly that many times with replacement from all genes,
what is then the probability of getting as much success as observed for the GOIs?
\itemize{
\item counts.DT: data.table with columns "id", "count" for number of observations for each of the ids
\item success_ids: ids that is considered a successful trial, e.g. ids that are annotated with a GO term being tested for.
\item n_total_ids: number of total ids so that number of background ids + success_ids == n_total_ids
}}
\usage{
binomial.p(counts.DT, success_ids, n_total_ids, log.p = F)
}
\description{
get the p-value from a binomial test where you have counts observed for some ids
and want to know if they are overrepresented among a set of "success" ids vs a background of all ids.
Example: For a set of genes of interest (GOI) each observed a number of times,
if we sample randomly that many times with replacement from all genes,
what is then the probability of getting as much success as observed for the GOIs?
\itemize{
\item counts.DT: data.table with columns "id", "count" for number of observations for each of the ids
\item success_ids: ids that is considered a successful trial, e.g. ids that are annotated with a GO term being tested for.
\item n_total_ids: number of total ids so that number of background ids + success_ids == n_total_ids
}
}
